Update volt-editable_text for 0.9
