+++ Gemfile
@@ -1,4 +1,4 @@
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in volt-editable-text.gemspec
+# Specify your gem's dependencies in volt-editable_text.gemspec
 gemspec
