--- app/editable-text/views/main/index.html
@@ -1,6 +1,0 @@
-<:Body>
-  {{if toggled}}
-    <input type="text" e-keypress="edit(event)" size="{{size}}" value="{{value}}" />
-  {{else}}
-    <span e-click="toggle_editing">{{value}}</span>
-  {{end}}
