+++ app/editable_text/views/main/index.html
@@ -1,0 +1,6 @@
+<:Body>
+  {{if toggled}}
+    <input type="text" e-keypress="edit(event)" e-focusout="blur" size="{{size}}" value="{{value}}" />
+  {{else}}
+    <span e-click="toggle_editing">{{value}}</span>
+  {{end}}
