+++ lib/volt/editable_text.rb
@@ -1,0 +1,6 @@
+module Volt
+  class EditableText
+    VERSION = '0.0.9'
+    # Your code goes here...
+  end
+end
